p_wave_detector/requirements.txt
kafka-python
obspy
numpy
tensorflow==2.10.0
influxdb
influxdb_client